pyproject.toml
[project]
name = "gripper-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <3.10"
dependencies = [
    "numpy>=2.0.2",
    "pybullet-mm>=3.2.7.post0",
    "google-generativeai>=0.8.0",
    "pillow>=10.0.0",
    "scipy>=1.13.0",
]
